Gawker
A fail, basically. They default to Latest Stories. The tab is
populated with Most Commented / Most Popular by a click in the
dropdown; the links are not present in the HTML so I guess there is an
ajax call. Not sure how much of that can be simulated in jsdom.